+++ src/api/chia/util/streamable.ts
@@ -1,6 +1,6 @@
 import {bytes, uint16} from "../types/_python_types_";
 
-export type VersionBlob = {
+export type VersionedBlob = {
   version: uint16
   blob: bytes
 };