+++ src/api/chia/wallet/puzzles/clawback/metadata.ts
@@ -1,5 +1,6 @@
 import {bool, uint16, uint64} from "../../../types/_python_types_";
 import {bytes32} from "../../../types/blockchain_format/sized_bytes";
+import {Marshall} from "../../../rpc/util";
 
 export type AutoClaimSettings = {
   enabled: bool;
@@ -7,6 +8,9 @@ export type AutoClaimSettings = {
   min_amount: uint64;
   batch_size: uint16;
 };
+
+export type AutoClaimSettingsCHIP0029 = AutoClaimSettings & Marshall;
+
 
 export type ClawbackMetadata = {
   time_lock: uint64;