+++ src/api/rpc/wallet/index.ts
@@ -35,10 +35,9 @@ import {TDaemon} from "../../../daemon/index";
 import {CoinRecord} from "../../chia/types/coin_record";
 import {SigningMode} from "../../chia/types/signing_mode";
 import {Balance} from "../../chia/wallet/wallet_node";
-import {AutoClaimSettings} from "../../chia/wallet/puzzles/clawback/metadata";
+import {AutoClaimSettings, AutoClaimSettingsCHIP0029} from "../../chia/wallet/puzzles/clawback/metadata";
 import {GetCoinRecords} from "../../chia/wallet/wallet_coin_store";
 import {WalletCoinRecordWithMetadata} from "../../chia/wallet/wallet_coin_record";
-import {VCRecord} from "../../chia/wallet/vc_wallet/vc_store";
 import {TransactionTypeFilter} from "../../chia/wallet/util/quality_filter";
 import {TXConfigLoader} from "../../chia/wallet/util/tx_config";
 import {ConditionValidTimes} from "../../chia/wallet/conditions";
@@ -46,44 +45,87 @@ import {DAOInfo, DAORules, ProposalInfo} from "../../chia/wallet/dao_wallet/dao_
 import {ParsedProposalSpend, ParsedProposalUpdate, ProposalState} from "../../chia/wallet/dao_wallet/dao_wallet";
 import {DLProof, VerifyProofResponse} from "../../chia/data_layer/data_layer_util";
 import {
-  GetNotifications,
-  GetNotificationsResponse,
-  GatherSigningInfo,
-  GatherSigningInfoCHIP0029,
-  GatherSigningInfoResponse,
-  GatherSigningInfoResponseCHIP0029,
+  AddKeyRequest,
+  AddKeyResponse,
   ApplySignatures,
   ApplySignaturesCHIP0029,
   ApplySignaturesResponse,
   ApplySignaturesResponseCHIP0029,
+  CheckDeleteKeyRequest,
+  CheckDeleteKeyResponse,
+  CombineCoins,
+  CombineCoinsCHIP0029,
+  CombineCoinsResponse,
+  CombineCoinsResponseCHIP0029,
+  DeleteKeyRequest,
+  ExecuteSigningInstructions,
+  ExecuteSigningInstructionsCHIP0029,
+  ExecuteSigningInstructionsResponse,
+  ExecuteSigningInstructionsResponseCHIP0029,
+  GatherSigningInfo,
+  GatherSigningInfoCHIP0029,
+  GatherSigningInfoResponse,
+  GatherSigningInfoResponseCHIP0029,
+  GenerateMnemonicResponse,
+  GetHeightInfo,
+  GetHeightInfoResponse,
+  GetLoggedInFingerprintResponse,
+  GetNotifications,
+  GetNotificationsResponse,
+  GetPrivateKeyRequest,
+  GetPrivateKeyResponse,
+  GetPublicKeysResponse,
+  GetSyncStatus,
+  GetSyncStatusResponse,
+  GetTimestampForHeightRequest,
+  GetTimestampForHeightRequestCHIP0029,
+  GetTimestampForHeightResponse,
+  LogIn,
+  LogInResponse,
+  PushTransactions,
+  PushTransactionsCHIP0029,
+  PushTransactionsResponse,
+  PushTransactionsResponseCHIP0029,
+  PushTX,
+  PushTXCHIP0029,
+  SetWalletResyncOnStartup,
+  SetWalletResyncOnStartupCHIP0029,
+  SplitCoins,
+  SplitCoinsCHIP0029,
+  SplitCoinsResponse,
+  SplitCoinsResponseCHIP0029,
   SubmitTransactions,
   SubmitTransactionsCHIP0029,
   SubmitTransactionsResponse,
   SubmitTransactionsResponseCHIP0029,
-  ExecuteSigningInstructionsCHIP0029,
-  ExecuteSigningInstructionsResponseCHIP0029,
-  ExecuteSigningInstructionsResponse,
-  ExecuteSigningInstructions,
-  SplitCoins,
-  SplitCoinsResponse,
-  SplitCoinsResponseCHIP0029,
-  CombineCoinsResponse,
-  CombineCoins,
-  CombineCoinsResponseCHIP0029,
-  LogIn,
-  LogInResponse,
-  GetLoggedInFingerprintResponse,
-  GetPublicKeysResponse,
-  GetPrivateKeyRequest,
-  GetPrivateKeyResponse,
-  GenerateMnemonicResponse,
-  AddKeyRequest,
-  AddKeyResponse,
-  DeleteKeyRequest, CheckDeleteKeyRequest, CheckDeleteKeyResponse,
+  VcAddProofs,
+  VcAddProofsCHIP0029,
+  VCGet,
+  VCGetCHIP0029,
+  VcGetList,
+  VcGetListCHIP0029,
+  VcGetListResponse,
+  VCGetProofsForRoot,
+  VCGetProofsForRootCHIP0029,
+  VCGetProofsForRootResponse,
+  VcGetResponse,
+  VCMint,
+  VCMintCHIP0029,
+  VcMintResponse,
+  VcMintResponseCHIP0029,
+  VcRevoke,
+  VcRevokeCHIP0029,
+  VcRevokeResponse,
+  VcRevokeResponseCHIP0029,
+  VcSpend,
+  VcSpendCHIP0029,
+  VcSpendResponse,
+  VcSpendResponseCHIP0029,
 } from "../../chia/rpc/wallet_request_types";
-import {TXEndpointRequest, TxeResp} from "../../chia/rpc/util";
 import {SigningResponse, UnsignedTransaction} from "../../chia/wallet/signer_protocol";
 import {WalletSpendBundle} from "../../chia/wallet/wallet_spend_bundle";
+import {TXEndpointRequest, TxeResp} from "../../chia/rpc/wallet_rpc_api";
+import {Marshall, MaybeMarshall} from "../../chia/rpc/util";
 
 export const chia_wallet_service = "chia_wallet";
 export type chia_wallet_service = typeof chia_wallet_service;
@@ -195,9 +237,7 @@ export async function delete_all_keys<T extends TRPCAgent | TDaemon>(agent: T){
 
 export const set_wallet_resync_on_startup_command = "set_wallet_resync_on_startup";
 export type set_wallet_resync_on_startup_command = typeof set_wallet_resync_on_startup_command;
-export type TSetWalletResyncOnStartupRequest = {
-  enable?: bool;
-};
+export type TSetWalletResyncOnStartupRequest = SetWalletResyncOnStartup | SetWalletResyncOnStartupCHIP0029;
 export type TSetWalletResyncOnStartupResponse = {
   success: True;
 };
@@ -211,14 +251,11 @@ export async function set_wallet_resync_on_startup<T extends TRPCAgent | TDaemon
 // # Wallet node
 export const get_sync_status_command = "get_sync_status";
 export type get_sync_status_command = typeof get_sync_status_command;
-export type TGetSyncStatusResponse = {
-  synced: bool;
-  syncing: bool;
-  genesis_initialized: bool;
-};
-export type WsGetSyncStatusMessage = GetMessageType<chia_wallet_service, get_sync_status_command, TGetSyncStatusResponse>;
-export async function get_sync_status<T extends TRPCAgent | TDaemon>(agent: T){
-  type R = ResType<T, TGetSyncStatusResponse, WsGetSyncStatusMessage>;
+export type TGetSyncStatus = GetSyncStatus;
+export type TGetSyncStatusResponse = GetSyncStatusResponse;
+export type WsGetSyncStatusMessage<R> = GetMessageType<chia_wallet_service, get_sync_status_command, R>;
+export async function get_sync_status<T extends TRPCAgent | TDaemon, D extends TGetSyncStatus|undefined=undefined>(agent: T, data?: D){
+  type R = ResType<T, GetSyncStatusResponse, WsGetSyncStatusMessage<GetSyncStatusResponse>>;
   return agent.sendMessage<R>(chia_wallet_service, get_sync_status_command);
 }
 
@@ -226,12 +263,11 @@ export async function get_sync_status<T extends TRPCAgent | TDaemon>(agent: T){
 
 export const get_height_info_command = "get_height_info";
 export type get_height_info_command = typeof get_height_info_command;
-export type TGetHeightInfoResponse = {
-  height: uint32;
-};
-export type WsGetHeightInfoMessage = GetMessageType<chia_wallet_service, get_height_info_command, TGetHeightInfoResponse>;
-export async function get_height_info<T extends TRPCAgent | TDaemon>(agent: T){
-  type R = ResType<T, TGetHeightInfoResponse, WsGetHeightInfoMessage>;
+export type TGetHeightInfoRequest = GetHeightInfo;
+export type TGetHeightInfoResponse = GetHeightInfoResponse;
+export type WsGetHeightInfoMessage<R> = GetMessageType<chia_wallet_service, get_height_info_command, R>;
+export async function get_height_info<T extends TRPCAgent | TDaemon, D extends TGetHeightInfoRequest|undefined=undefined>(agent: T, data?: D){
+  type R = ResType<T, GetHeightInfoResponse, WsGetHeightInfoMessage<GetHeightInfoResponse>>;
   return agent.sendMessage<R>(chia_wallet_service, get_height_info_command);
 }
 
@@ -239,9 +275,7 @@ export async function get_height_info<T extends TRPCAgent | TDaemon>(agent: T){
 
 export const push_tx_command = "push_tx";
 export type push_tx_command = typeof push_tx_command;
-export type TPushTxRequest = {
-  spend_bundle: str; // SpendBundle serialized to hex string 
-};
+export type TPushTxRequest = PushTX | PushTXCHIP0029;
 export type TPushTxResponse = Record<string, never>;
 export type WsPushTxMessageOfWallet = GetMessageType<chia_wallet_service, push_tx_command, TPushTxResponse>;
 export async function push_tx<T extends TRPCAgent | TDaemon>(agent: T, data: TPushTxRequest){
@@ -253,14 +287,12 @@ export async function push_tx<T extends TRPCAgent | TDaemon>(agent: T, data: TPu
 
 export const push_transactions_command = "push_transactions";
 export type push_transactions_command = typeof push_transactions_command;
-export type TPushTransactionsRequest = {
-  transactions: Array<TransactionRecordConvenience | str>; // TransactionRecord or hex-serialized string
-  fee?: uint64;
-} & TXEndpointRequest;
-export type TPushTransactionsResponse = Record<string, never>;
-export type WsPushTransactionsMessage = GetMessageType<chia_wallet_service, push_transactions_command, TPushTransactionsResponse>;
-export async function push_transactions<T extends TRPCAgent | TDaemon>(agent: T, data: TPushTransactionsRequest){
-  type R = ResType<T, TPushTransactionsResponse, WsPushTransactionsMessage>;
+export type TPushTransactionsRequest = PushTransactions | PushTransactionsCHIP0029;
+export type TPushTransactionsResponse = PushTransactionsResponse | PushTransactionsResponseCHIP0029;
+export type WsPushTransactionsMessage<R> = GetMessageType<chia_wallet_service, push_transactions_command, R>;
+export async function push_transactions<T extends TRPCAgent | TDaemon, D extends PushTransactions|PushTransactionsCHIP0029>(agent: T, data: D){
+  type MM = MaybeMarshall<D, PushTransactionsResponse, PushTransactionsResponseCHIP0029>;
+  type R = ResType<T, MM, WsPushTransactionsMessage<MM>>;
   return agent.sendMessage<R>(chia_wallet_service, push_transactions_command, data);
 }
 
@@ -268,35 +300,37 @@ export async function push_transactions<T extends TRPCAgent | TDaemon>(agent: T,
 
 export const get_timestamp_for_height_command = "get_timestamp_for_height";
 export type get_timestamp_for_height_command = typeof get_timestamp_for_height_command;
-export type TGetTimestampForHeightResponse = {
-  timestamp: uint64;
-};
-export type WsGetTimestampForHeightMessage = GetMessageType<chia_wallet_service, get_timestamp_for_height_command, TGetTimestampForHeightResponse>;
-export async function get_timestamp_for_height<T extends TRPCAgent | TDaemon>(agent: T) {
-  type R = ResType<T, TGetTimestampForHeightResponse, WsGetTimestampForHeightMessage>;
-  return agent.sendMessage<R>(chia_wallet_service, get_timestamp_for_height_command);
+export type TGetTimestampForHeightRequest = GetTimestampForHeightRequest | GetTimestampForHeightRequestCHIP0029;
+export type TGetTimestampForHeightResponse = GetTimestampForHeightResponse;
+export type WsGetTimestampForHeightMessage<R> = GetMessageType<chia_wallet_service, get_timestamp_for_height_command, R>;
+export async function get_timestamp_for_height<T extends TRPCAgent | TDaemon, D extends TGetTimestampForHeightRequest>(agent: T, params: D) {
+  type R = ResType<T, GetTimestampForHeightResponse, WsGetTimestampForHeightMessage<GetTimestampForHeightResponse>>;
+  return agent.sendMessage<R>(chia_wallet_service, get_timestamp_for_height_command, params);
 }
 
 
 
 export const set_auto_claim_command = "set_auto_claim";
 export type set_auto_claim_command = typeof set_auto_claim_command;
-export type TSetAutoClaimRequest = AutoClaimSettings;
-export type TSetAutoClaimResponse = AutoClaimSettings;
-export type WsSetAutoClaimMessage = GetMessageType<chia_wallet_service, set_auto_claim_command, TSetAutoClaimResponse>;
-export async function set_auto_claim<T extends TRPCAgent | TDaemon>(agent: T, data: TSetAutoClaimRequest) {
-  type R = ResType<T, TSetAutoClaimResponse, WsSetAutoClaimMessage>;
+export type TSetAutoClaimRequest = AutoClaimSettings | (AutoClaimSettingsCHIP0029 & Marshall);
+export type TSetAutoClaimResponse = AutoClaimSettings | AutoClaimSettingsCHIP0029;
+export type WsSetAutoClaimMessage<R> = GetMessageType<chia_wallet_service, set_auto_claim_command, R>;
+export async function set_auto_claim<T extends TRPCAgent | TDaemon, D extends TSetAutoClaimRequest>(agent: T, data: D) {
+  type MM = MaybeMarshall<D, AutoClaimSettings, AutoClaimSettingsCHIP0029>;
+  type R = ResType<T, MM, WsSetAutoClaimMessage<MM>>;
   return agent.sendMessage<R>(chia_wallet_service, set_auto_claim_command, data);
 }
 
 
 export const get_auto_claim_command = "get_auto_claim";
 export type get_auto_claim_command = typeof get_auto_claim_command;
-export type TGetAutoClaimResponse = AutoClaimSettings;
-export type WsGetAutoClaimMessage = GetMessageType<chia_wallet_service, get_auto_claim_command, TGetAutoClaimResponse>;
-export async function get_auto_claim<T extends TRPCAgent | TDaemon>(agent: T) {
-  type R = ResType<T, TGetAutoClaimResponse, WsGetAutoClaimMessage>;
-  return agent.sendMessage<R>(chia_wallet_service, get_auto_claim_command);
+export type TGetAutoClaimRequest = Marshall;
+export type TGetAutoClaimResponse = AutoClaimSettings | AutoClaimSettingsCHIP0029;
+export type WsGetAutoClaimMessage<R> = GetMessageType<chia_wallet_service, get_auto_claim_command, R>;
+export async function get_auto_claim<T extends TRPCAgent | TDaemon, D extends TGetAutoClaimRequest|undefined=undefined>(agent: T, params?: D) {
+  type MM = MaybeMarshall<D, AutoClaimSettings, AutoClaimSettingsCHIP0029>;
+  type R = ResType<T, MM, WsGetAutoClaimMessage<MM>>;
+  return agent.sendMessage<R>(chia_wallet_service, get_auto_claim_command, params);
 }
 
 
@@ -961,7 +995,7 @@ export async function get_transaction_memo<T extends TRPCAgent | TDaemon>(agent:
 
 export const split_coins_command = "split_coins";
 export type split_coins_command = typeof split_coins_command;
-export type TSplitCoinsRequest = SplitCoins & TXEndpointRequest;
+export type TSplitCoinsRequest = SplitCoins | SplitCoinsCHIP0029;
 export type TSplitCoinsResponse = SplitCoinsResponse | SplitCoinsResponseCHIP0029;
 export type WsSplitCoinsMessage<R> = GetMessageType<chia_wallet_service, split_coins_command, R>;
 export async function split_coins<T extends TRPCAgent | TDaemon, D extends TSplitCoinsRequest>(
@@ -978,7 +1012,7 @@ export async function split_coins<T extends TRPCAgent | TDaemon, D extends TSpli
 
 export const combine_coins_command = "combine_coins";
 export type combine_coins_command = typeof combine_coins_command;
-export type TCombineCoinsRequest = CombineCoins & TXEndpointRequest;
+export type TCombineCoinsRequest = CombineCoins & CombineCoinsCHIP0029;
 export type TCombineCoinsResponse = CombineCoinsResponse | CombineCoinsResponseCHIP0029;
 export type WsCombineCoinsMessage<R> = GetMessageType<chia_wallet_service, combine_coins_command, R>;
 export async function combine_coins<T extends TRPCAgent | TDaemon, D extends TCombineCoinsRequest>(
@@ -2863,36 +2897,23 @@ export async function dl_verify_proof<T extends TRPCAgent | TDaemon>(agent: T, d
 }
 
 
-export type VCMint = {
-  did_id: str;
-  target_address: Optional<str>;
-  fee: uint64;
-};
 export const vc_mint_command = "vc_mint";
 export type vc_mint_command = typeof vc_mint_command;
-export type TVcMintRequest = VCMint & TXEndpointRequest;
-export type TVcMintResponse = {
-  vc_record: VCRecord;
-  transactions: TransactionRecordConvenience[];
-  signing_responses?: str[];
-};
+export type TVcMintRequest = VCMint | VCMintCHIP0029;
+export type TVcMintResponse = VcMintResponse | VcMintResponseCHIP0029;
 export type WsVcMintMessage<R> = GetMessageType<chia_wallet_service, vc_mint_command, R>;
 export async function vc_mint<T extends TRPCAgent | TDaemon, D extends TVcMintRequest>(agent: T, data: D) {
-  type TER = TxeResp<D, TVcMintResponse>;
+  type MM = MaybeMarshall<D, VcMintResponse, VcMintResponseCHIP0029>;
+  type TER = TxeResp<D, MM>;
   type R = ResType<T, TER, WsVcMintMessage<TER>>;
   return agent.sendMessage<R>(chia_wallet_service, vc_mint_command, data);
 }
 
 
-export type VCGet = {
-  vc_id: bytes32;
-};
 export const vc_get_command = "vc_get";
 export type vc_get_command = typeof vc_get_command;
-export type TVcGetRequest = VCGet;
-export type TVcGetResponse = {
-  vc_record: VCRecord | None;
-};
+export type TVcGetRequest = VCGet | VCGetCHIP0029;
+export type TVcGetResponse = VcGetResponse;
 export type WsVcGetMessage = GetMessageType<chia_wallet_service, vc_get_command, TVcGetResponse>;
 export async function vc_get<T extends TRPCAgent | TDaemon>(agent: T, data: TVcGetRequest) {
   type R = ResType<T, TVcGetResponse, WsVcGetMessage>;
@@ -2900,17 +2921,10 @@ export async function vc_get<T extends TRPCAgent | TDaemon>(agent: T, data: TVcG
 }
 
 
-export type VcGetList = {
-  start: uint32;
-  end: uint32;
-};
 export const vc_get_list_command = "vc_get_list";
 export type vc_get_list_command = typeof vc_get_list_command;
-export type TVcGetListRequest = VcGetList;
-export type TVcGetListResponse = {
-  vc_records: Array<VCRecord & {coin_id: str;}>;
-  proofs: Record<str, Record<str, str> | None>;
-};
+export type TVcGetListRequest = VcGetList | VcGetListCHIP0029;
+export type TVcGetListResponse = VcGetListResponse;
 export type WsVcGetListMessage = GetMessageType<chia_wallet_service, vc_get_list_command, TVcGetListResponse>;
 export async function vc_get_list<T extends TRPCAgent | TDaemon>(agent: T, data: TVcGetListRequest) {
   type R = ResType<T, TVcGetListResponse, WsVcGetListMessage>;
@@ -2918,23 +2932,14 @@ export async function vc_get_list<T extends TRPCAgent | TDaemon>(agent: T, data:
 }
 
 
-export type VcSpend = {
-  vc_id: bytes32;
-  new_puzhash: Optional<bytes32>;
-  new_proof_hash: Optional<bytes32>;
-  provider_inner_puzhash: Optional<bytes32>;
-  fee: uint64;
-};
 export const vc_spend_command = "vc_spend";
 export type vc_spend_command = typeof vc_spend_command;
-export type TVcSpendRequest = VcSpend & TXEndpointRequest;
-export type TVcSpendResponse = {
-  transactions: TransactionRecordConvenience[];
-  signing_responses?: str[];
-};
+export type TVcSpendRequest = VcSpend | VcSpendCHIP0029;
+export type TVcSpendResponse = VcSpendResponse | VcSpendResponseCHIP0029;
 export type WsVcSpendMessage<R> = GetMessageType<chia_wallet_service, vc_spend_command, R>;
 export async function vc_spend<T extends TRPCAgent | TDaemon, D extends TVcSpendRequest>(agent: T, data: D) {
-  type TER = TxeResp<D, TVcSpendResponse>;
+  type MM = MaybeMarshall<D, VcSpendResponse, VcSpendResponseCHIP0029>;
+  type TER = TxeResp<D, MM>;
   type R = ResType<T, TER, WsVcSpendMessage<TER>>;
   return agent.sendMessage<R>(chia_wallet_service, vc_spend_command, data);
 }
@@ -2942,11 +2947,7 @@ export async function vc_spend<T extends TRPCAgent | TDaemon, D extends TVcSpend
 
 export const vc_add_proofs_command = "vc_add_proofs";
 export type vc_add_proofs_command = typeof vc_add_proofs_command;
-export type TVcAddProofsRequest = {
-  proofs: {
-    key_value_pairs: Record<str, str>;
-  };
-};
+export type TVcAddProofsRequest = VcAddProofs | VcAddProofsCHIP0029;
 export type TVcAddProofsResponse = Record<string, never>;
 export type WsVcAddProofsMessage = GetMessageType<chia_wallet_service, vc_add_proofs_command, TVcAddProofsResponse>;
 export async function vc_add_proofs<T extends TRPCAgent | TDaemon>(agent: T, data: TVcAddProofsRequest) {
@@ -2955,15 +2956,10 @@ export async function vc_add_proofs<T extends TRPCAgent | TDaemon>(agent: T, dat
 }
 
 
-export type VCGetProofsForRoot = {
-  root: bytes32;
-};
 export const vc_get_proofs_for_root_command = "vc_get_proofs_for_root";
 export type vc_get_proofs_for_root_command = typeof vc_get_proofs_for_root_command;
-export type TVcGetProofsForRootRequest = VCGetProofsForRoot;
-export type TVcGetProofsForRootResponse = {
-  proofs: Record<str, str>;
-};
+export type TVcGetProofsForRootRequest = VCGetProofsForRoot | VCGetProofsForRootCHIP0029;
+export type TVcGetProofsForRootResponse = VCGetProofsForRootResponse;
 export type WsVcGetProofsForRootMessage = GetMessageType<chia_wallet_service, vc_get_proofs_for_root_command, TVcGetProofsForRootResponse>;
 export async function vc_get_proofs_for_root<T extends TRPCAgent | TDaemon>(agent: T, data: TVcGetProofsForRootRequest) {
   type R = ResType<T, TVcGetProofsForRootResponse, WsVcGetProofsForRootMessage>;
@@ -2971,20 +2967,14 @@ export async function vc_get_proofs_for_root<T extends TRPCAgent | TDaemon>(agen
 }
 
 
-export type VcRevoke = {
-  vc_parent_id: bytes32;
-  fee: uint64;
-};
 export const vc_revoke_command = "vc_revoke";
 export type vc_revoke_command = typeof vc_revoke_command;
-export type TVcRevokeRequest = VcRevoke & TXEndpointRequest;
-export type TVcRevokeResponse = {
-  transactions: TransactionRecordConvenience[];
-  signing_responses?: str[];
-};
+export type TVcRevokeRequest = VcRevoke | VcRevokeCHIP0029;
+export type TVcRevokeResponse = VcRevokeResponse | VcRevokeResponseCHIP0029;
 export type WsVcRevokeMessage<R> = GetMessageType<chia_wallet_service, vc_revoke_command, R>;
 export async function vc_revoke<T extends TRPCAgent | TDaemon, D extends TVcRevokeRequest>(agent: T, data: D) {
-  type TER = TxeResp<D, TVcRevokeResponse>;
+  type MM = MaybeMarshall<D, VcRevokeResponse, VcRevokeResponseCHIP0029>;
+  type TER = TxeResp<D, MM>;
   type R = ResType<T, TER, WsVcRevokeMessage<TER>>;
   return agent.sendMessage<R>(chia_wallet_service, vc_revoke_command, data);
 }